Allow to add agents
Dialog related animations on iPhone not entirely smooth. There's a little jank moving dialog up when keyboard animates in, when animating the keyboard down, and the list stutters slightly when adding items.
It's ok-ish, but noticeably different to the native iOS UI.
The UI currently consists basically of just text labels, and there are unlikely any other app specific reasons for this performance.