Revert "Extend compose root view fully to screen bottom on iOS"
This reverts commit 350ff823172b9c89d57576757f65ffe2fafe92dc.

Looks fine if the keyboard is closed, but when showing it the input is pushed up all the way up to the top of the screen.
Extenting the root view to the bottom is not really needed (it was just an experimental tweak to color the home indicator area the same as the input), so reverting.