Integrate Hugging Face interface providers (#5)
* Integrate huggingface inference provider

* Move dialog to separate file

* Enter the access token in the UI

* Store access token in preferences

This adds general support for preferences

Note: not working on web, androidx.datastore currently works on all platforms except for web
will be fixed in a separate commit

* Allow to use prefs in web

Includes a large refactoring to be able to share code (in this case the default implementation of preferences) for non-web targets. This required to add a new source set ("nonWebMain") to the build file.

* Improve error handling