Implement spaces
A space is where the user interacts with an agent.
When an agent is created, a space referencing this agent is automatically created. This could be decoupled in the future.
A space currently can have only 1 agent. An agent can as per database schema appear in multiple spaces, but the UI currently doesn't allow this.